chore: Move IDE Sidebar into ADS (#38641)
