+++ app/client/packages/design-system/ads/src/Templates/Sidebar/SidebarButton/SidebarButton.types.ts
@@ -1,0 +1,12 @@
+import type { Condition } from "../enums";
+
+export interface SidebarButtonProps {
+  title?: string;
+  testId: string;
+  selected: boolean;
+  icon: string;
+  onClick: (urlSuffix: string) => void;
+  urlSuffix: string;
+  tooltip?: string;
+  condition?: Condition;
+}
