+++ app/client/packages/design-system/ads/src/Templates/Sidebar/SidebarButton/constants.ts
@@ -1,0 +1,14 @@
+import { Condition } from "../enums";
+
+export const ConditionConfig: Record<
+  Condition,
+  { icon: string; color: string }
+> = {
+  [Condition.Warn]: {
+    icon: "warning",
+    color: "#ffe283",
+  },
+  // TODO add this information for further conditions
+  // Error: { color: "", icon: "" },
+  // Success: { color: "", icon: "" },
+};
