+++ app/client/packages/design-system/ads/src/Templates/Sidebar/SidebarButton/index.ts
@@ -1,0 +1,1 @@
+export { SidebarButton } from "./SidebarButton";
