+++ app/client/packages/design-system/ads/src/Templates/Sidebar/enums.ts
@@ -1,0 +1,5 @@
+export enum Condition {
+  Warn = "Warn",
+  // Error = "Error",
+  // Success = "Success",
+}
