+++ app/client/packages/design-system/ads/src/Templates/Sidebar/index.tsx
@@ -1,0 +1,3 @@
+export { IDESidebar } from "./Sidebar";
+export type { IDESidebarButton } from "./types";
+export { Condition } from "./enums";
