+++ app/client/packages/design-system/ads/src/Templates/index.ts
@@ -1,2 +1,3 @@
 export * from "./IDEHeader";
 export * from "./EntityExplorer";
+export * from "./Sidebar";
