--- app/client/src/IDE/Components/Sidebar/SidebarButton/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./SidebarButton";
