--- app/client/src/IDE/Components/Sidebar/index.tsx
@@ -1,2 +1,0 @@
-export { default } from "./Sidebar";
-export type { IDESidebarButton } from "./Sidebar";
